+++ Examples/ARC Test/Classes/iRateAppDelegate.m
@@ -26,7 +26,8 @@
     //but we need to test with an app that's actually on the store
 	[iRate sharedInstance].appStoreID = 355313284;
     [iRate sharedInstance].applicationBundleID = @"com.charcoaldesign.rainbowblocks";
-	
+    [iRate sharedInstance].onlyPromptIfLatestVersion = NO;
+
     //enable debug mode
     [iRate sharedInstance].debug = YES;
 }