+++ Examples/Mac Demo/iRateMacAppDelegate.m
@@ -20,7 +20,8 @@
     //as it is picked up automatically from your Info.plist file
     //but we want to test with an app that's actually on the store
     [iRate sharedInstance].applicationBundleID = @"com.charcoaldesign.RainbowBlocksLite";
-	
+    [iRate sharedInstance].onlyPromptIfLatestVersion = NO;
+
     //enable debug mode
     [iRate sharedInstance].debug = YES;
 }